PMM-6542 move function to right files
