+++ pmm-app/src/pmm-add-instance-app-panel/components/AddRemoteInstance/AddRemoteInstance.tools.tsx
@@ -1,10 +1,5 @@
-import React, {useState} from 'react';
-import {InstanceData, TrackingOptions} from './AddRemoteInstance.types';
-import {RadioButtonGroup} from '../../../shared/components/Form/Radio/RadioButtonGroup';
-import {trackingOptions} from './AddRemoteInstance.constants';
-import {Messages} from './AddRemoteInstance.messages';
-import {CheckboxField} from '../../../shared/components/Form';
-import Validators from "../../../shared/components/helpers/validators";
+import Validators from 'shared/components/helpers/validators';
+import { InstanceData } from './AddRemoteInstance.types';
 
 export const extractCredentials = (credentials) => ({
   service_name: !credentials.isRDS ? credentials.address : credentials.instance_id,
@@ -17,6 +12,7 @@ export const extractCredentials = (credentials) => ({
   instance_id: credentials.instance_id,
   az: credentials.az,
 });
+
 export const getInstanceData = (instanceType, credentials) => {
   const instance: InstanceData = {};
 
@@ -44,62 +40,6 @@ export const getInstanceData = (instanceType, credentials) => {
   }
 
   return instance;
-};
-export const PostgreSQLAdditionalOptions = ({ mutators }) => {
-  const [trackingType, setTrackingType] = useState<string>(TrackingOptions.none);
-
-  return (
-    <>
-      <RadioButtonGroup
-        options={trackingOptions}
-        selected={trackingType}
-        name="tracking"
-        dataQa="tracking-options-radio-button-group"
-        onChange={(value) => {
-          mutators.changePGTracking(value);
-          setTrackingType(value);
-        }}
-      />
-      <span className="description">{Messages.form.labels.trackingOptions}</span>
-    </>
-  );
-};
-export const getAdditionalOptions = (type, remoteInstanceCredentials, mutators) => {
-  switch (type) {
-    case 'PostgreSQL':
-      return <PostgreSQLAdditionalOptions mutators={mutators} />;
-    case 'MySQL':
-      if (remoteInstanceCredentials.isRDS) {
-        return (
-          <>
-            <CheckboxField label="Use performance schema" name="qan_mysql_perfschema" />
-            <span className="description" />
-
-            <CheckboxField label="Disable Basic Metrics" name="disable_basic_metrics" />
-            <span className="description" />
-
-            <CheckboxField label="Disable Enhanced Metrics" name="disable_enhanced_metrics" />
-            <span className="description" />
-          </>
-        );
-      }
-
-      return (
-        <>
-          <CheckboxField label="Use performance schema" name="qan_mysql_perfschema" />
-          <span className="description" />
-        </>
-      );
-    case 'MongoDB':
-      return (
-        <>
-          <CheckboxField label="Use QAN MongoDB Profiler" name="qan_mongodb_profiler" />
-          <span className="description" />
-        </>
-      );
-    default:
-      return null;
-  }
 };
 export const validateInstanceForm = (values) => {
   const errors = {} as any;
