+++ pmm-app/src/pmm-add-instance-app-panel/components/AddRemoteInstance/AddRemoteInstance.tsx
@@ -1,19 +1,12 @@
 import React, { useState } from 'react';
 import { Form as FormFinal } from 'react-final-form';
-import {
-  CheckboxField,
-  PasswordInputField,
-  StepProgress,
-  TextareaInputField,
-  TextInputField,
-  validators,
-} from '@percona/platform-core';
-import AddRemoteInstanceService from './AddRemoteInstance.service';
-import { TrackingOptions } from './AddRemoteInstance.types';
-import { getAdditionalOptions, getInstanceData, validateInstanceForm } from './AddRemoteInstance.tools';
+import { StepProgress } from '@percona/platform-core';
 import { useTheme } from '@grafana/ui';
+import AddRemoteInstanceService, { toPayload } from './AddRemoteInstance.service';
+import { getInstanceData, validateInstanceForm } from './AddRemoteInstance.tools';
 import { getStyles } from './AddRemoteInstance.styles';
 import { Messages } from './AddRemoteInstance.messages';
+import { AdditionalOptionsFormPart, LabelsFormPart, MainDetailsFormPart } from './FormParts';
 
 const AddRemoteInstance = (props) => {
   const theme = useTheme();
@@ -22,6 +15,7 @@ const AddRemoteInstance = (props) => {
   const {
     instance: { type, credentials },
   } = props;
+
   const { instanceType, remoteInstanceCredentials, discoverName } = getInstanceData(type, credentials);
   const [loading, setLoading] = useState<boolean>(false);
   const initialValues = { ...remoteInstanceCredentials };
@@ -38,171 +32,49 @@ const AddRemoteInstance = (props) => {
     const currentUrl = `${window.parent.location}`;
     const newURL = `${currentUrl.split('/graph/d/').shift()}/graph/d/pmm-inventory/`;
 
-    const data = { ...values };
-
-    if (values.custom_labels) {
-      data.custom_labels = data.custom_labels
-        .split(/[\n\s]/)
-        .filter(Boolean)
-        .reduce((acc, val) => {
-          const [key, value] = val.split(':');
-
-          acc[key] = value;
-
-          return acc;
-        }, {});
-    }
-
-    switch (data.tracking) {
-      case TrackingOptions.pgStatements:
-        data.qan_postgresql_pgstatements_agent = true;
-        break;
-      case TrackingOptions.pgMonitor:
-        data.qan_postgresql_pgstatmonitor_agent = true;
-        break;
-      default:
-        delete data.tracking;
-    }
-
-    delete data.tracking;
-
-    if (!data.service_name) {
-      data.service_name = data.address;
-    }
-
-    if (data.add_node === undefined) {
-      data.add_node = {
-        node_name: data.service_name,
-        node_type: 'REMOTE_NODE',
-      };
-    }
-
-    if (discoverName) {
-      data.engine = discoverName;
-    }
-
-    if (data.pmm_agent_id === undefined || data.pmm_agent_id === '') {
-      data.pmm_agent_id = 'pmm-server'; // set default value for pmm agent id
-    }
-
-    setLoading(true);
+    const data = toPayload(values, discoverName);
 
     try {
+      setLoading(true);
+
       if (values.isRDS) {
-        data.rds_exporter = true;
         await AddRemoteInstanceService.addRDS(data);
       } else {
-        // remove rds fields from data
         await AddRemoteInstanceService.addRemote(instanceType, data);
       }
 
-      setLoading(false);
       window.location.assign(newURL);
     } catch (e) {
+      console.error(e);
+    } finally {
       setLoading(false);
     }
   };
 
-  const getSteps = (form) => {
-    return [
-      {
-        title: Messages.form.titles.mainDetails,
-        fields: ['address', 'port', 'username', 'password'],
-        render: () => (
-          <div className="" style={{ width: '50%' }}>
-            <TextInputField
-              name="address"
-              label={Messages.form.labels.mainDetails.address}
-              placeholder={'Hostname'}
-              validators={[validators.required]}
-              disabled={remoteInstanceCredentials.isRDS}
-            />
-            <TextInputField
-              name="service_name"
-              label={Messages.form.labels.mainDetails.serviceName}
-              placeholder={'Service name (default: Hostname)'}
-            />
-            <TextInputField
-              name="port"
-              label={Messages.form.labels.mainDetails.port}
-              placeholder={`Port (default: ${remoteInstanceCredentials.port} )`}
-              validators={[validators.required]}
-            />
-            <TextInputField
-              name="username"
-              label={Messages.form.labels.mainDetails.username}
-              placeholder={'Username'}
-              validators={[validators.required]}
-              disabled={remoteInstanceCredentials.isRDS}
-            />
-            <PasswordInputField
-              name="password"
-              label={Messages.form.labels.mainDetails.password}
-              placeholder={'Password'}
-              validators={[validators.required]}
-            />
-          </div>
-        ),
-      },
-      {
-        title: Messages.form.titles.labels,
-        fields: ['topology', 'nodes', 'databaseType'],
-        render: () => (
-          <div className="" style={{ width: '50%' }}>
-            <TextInputField
-              name="environment"
-              label={Messages.form.labels.labels.environment}
-              placeholder={'Environment'}
-            />
-            <TextInputField name="region" label={Messages.form.labels.labels.region} placeholder={'Region'} />
-            <TextInputField
-              name="az"
-              label={Messages.form.labels.labels.az}
-              placeholder={'Availability Zone'}
-            />
-            <TextInputField
-              name="replication_set"
-              label={Messages.form.labels.labels.replicationSet}
-              placeholder={'Replication set'}
-            />
-            <TextInputField
-              name="cluster"
-              label={Messages.form.labels.labels.cluster}
-              placeholder={'Cluster'}
-            />
-            <TextareaInputField
-              name="custom_labels"
-              label={Messages.form.labels.labels.customLabels}
-              placeholder="Custom labels
-      Format:
-      key1:value1
-      key2:value2"
-            />
-          </div>
-        ),
-      },
-      {
-        title: Messages.form.titles.additionalOptions,
-        fields: [],
-        render: () => (
-          <div className="" style={{ width: '50%' }}>
-            <div className="add-instance-panel" style={{ alignItems: 'flex-start' }}>
-              <CheckboxField label="Skip connection check" name="skip_connection_check" />
-              <CheckboxField label="Use TLS for database connections" name="tls" />
-              <CheckboxField label="Skip TLS certificate and hostname validation" name="tls_skip_verify" />
-              {getAdditionalOptions(instanceType, remoteInstanceCredentials, form.mutators)}
-            </div>
-
-            <div>
-              <button type="submit" className="button button--dark" id="addInstance" disabled={loading}>
-                Add service
-              </button>
-            </div>
-          </div>
-        ),
-      },
-    ];
-  };
+  const getSteps = (form) => [
+    {
+      title: Messages.form.titles.mainDetails,
+      fields: ['address', 'port', 'username', 'password'],
+      render: () => <MainDetailsFormPart remoteInstanceCredentials={remoteInstanceCredentials} />,
+    },
+    {
+      title: Messages.form.titles.labels,
+      fields: ['topology', 'nodes', 'databaseType'],
+      render: () => <LabelsFormPart />,
+    },
+    {
+      title: Messages.form.titles.additionalOptions,
+      fields: [],
+      render: () => (
+        <AdditionalOptionsFormPart
+          remoteInstanceCredentials={remoteInstanceCredentials}
+          loading={loading}
+          form={form}
+          instanceType={instanceType}
+        />
+      ),
+    },
+  ];
 
   return (
     <div className={styles.formWrapper}>
